+++ tencentcloud/ie/v20200304/errors.go
@@ -68,6 +68,9 @@ const (
 	// 参数错误。
 	INVALIDPARAMETER = "InvalidParameter"
 
+	// 不支持直播流。
+	INVALIDPARAMETER_LIVESTREAMNOTSUPPORT = "InvalidParameter.LiveStreamNotSupport"
+
 	// 参数取值错误。
 	INVALIDPARAMETERVALUE = "InvalidParameterValue"